Move `crypto::MultisigAddress` to separate file
