+++ src/crypto/address.rs
@@ -1,4 +1,5 @@
 use super::{ChecksumAlg, BASE32_NOPAD, CHECKSUM_LEN, HASH_LEN};
+use crate::models::Ed25519PublicKey;
 use sha2::Digest;
 
 /// Public key address
@@ -36,6 +37,54 @@ impl Address {
         let checksum = &hashed[(HASH_LEN - CHECKSUM_LEN)..];
         let checksum_address = [&self.0, checksum].concat();
         BASE32_NOPAD.encode(&checksum_address)
+    }
+}
+
+/// Convenience struct for handling multisig public identities
+#[derive(Debug, Clone)]
+pub struct MultisigAddress {
+    /// the version of this multisig
+    pub version: u8,
+    /// how many signatures are needed to fully sign as this address
+    pub threshold: u8,
+    /// ordered list of public keys that could potentially sign a message
+    pub public_keys: Vec<Ed25519PublicKey>,
+}
+
+impl MultisigAddress {
+    pub fn new(
+        version: u8,
+        threshold: u8,
+        addresses: &[Address],
+    ) -> Result<MultisigAddress, String> {
+        if version != 1 {
+            Err("Unknown msig version".to_string())
+        } else if threshold == 0 || addresses.is_empty() || threshold > addresses.len() as u8 {
+            Err("Invalid threshold".to_string())
+        } else {
+            Ok(MultisigAddress {
+                version,
+                threshold,
+                public_keys: addresses
+                    .iter()
+                    .map(|address| Ed25519PublicKey(address.0))
+                    .collect(),
+            })
+        }
+    }
+
+    /// Generates a checksum from the contained public keys usable as an address
+    pub fn address(&self) -> Address {
+        let mut buf = b"MultisigAddr".to_vec();
+        buf.push(self.version);
+        buf.push(self.threshold);
+        for key in &self.public_keys {
+            buf.extend_from_slice(&key.0);
+        }
+        let hashed = ChecksumAlg::digest(&buf);
+        let mut bytes = [0; HASH_LEN];
+        bytes.copy_from_slice(&hashed);
+        Address::new(bytes)
     }
 }
